refactor: split proxy logic in proxy.rs
